AmazonWebsite/requirements.txt
blinker==1.6.2
certifi==2024.7.4
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.7
importlib-metadata==6.7.0
intasend-python==1.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.32.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.2.2
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.19.1
flask-migrate
